main/requirements.txt
# local package
-e .

# external requirements
awscli
flake8
python-dotenv>=0.5.1
hydra-core
torchvision
matplotlib
numpy
omegaconf
pytorch_lightning
math
pytorch
statistics
matplotlib
os
torchvision
random
copy
scipy
PIL
typing
torchvision
csv
cv2
scikit-learn
sys